Block helpers mutate in place rather than return
Previously, invoking a block helper would always return a string or
element that would be inserted into the parent fragment.

This model assumed that the root-most template was always being rendered
exactly once.

With the introduction of re-rendering, block helpers need to handle this
case as well. Specifically, instead of just creating rendering a new
element each time, they should populate a provided morph with the
contents they want to display.

As an optimization, you can also save off the result of the previous
render and `rerender()` it.

This change is necessary to support re-rendering, but we consider it a
low-level API. We intended to create a high-level API for authors to use
that handles the most frequent rendering and re-rendering patterns.

In particular, the test titled “Templates with block helpers -
re-rendering” shows a low-level pattern that can be used to support
efficient re-rendering in general, and which should be easy to extract
into something higher level once we have made sure we have handled all
of the necessary cases.