Always revalidate child nodes for all statements
Previously, we assumed that `inline`, `content`, and other non-block
statements could never have child nodes. However, keywords can change
the meaning of one of these statements so that they actually behave like
a block (and have child nodes).

In those situations, we need to make sure that we continue revalidating
the child node hierarchy, even though no template was passed to the
hook.