Use the templateId instead of template identity
This allows us to use a unified code path between rehydration and normal
rendering.

We also discovered that we don’t use `lastYielded.self` (it was an
accidental holdover from the code that was refactored into lastYielded),
so we removed it (along with a vestigial `shadowTemplate`, removed in a
previous commit).