Add temporary workaround for SimpleDOM
The current parseHTML implementation requires some DOM APIs not
implemented in SimpleDOM. This commit adds a temporary workaround
to support using SimpleDOM on the server.

In the future, we will move towards using `insertAdjacentHTML` and
`innerHTML=` as the main APIs and remove the need of `parseHTML`
altogether.

This commit also reverts a previous commit that introduced the
`setMorphHTML` method, which does exactly what `Morph#setHTML`
does.