[BUGFIX beta] Range redirection
When an `range` calls `handleRedirect` it sets the `params` to an array
containing its value. This causes problems with contextual components called
with dot syntax.

If a contextual component is called with dot syntax and any kind of arguments
(either `params` and `attrs`) it gets compiled to `inline`. But when called
without any of them (`{{contextual.my-component}}`), it gets compiled to
a `range`.

The `range` calls to `handleRedirect` but sets `params` to be `[value]` instead
of an empty array, causing `positionalParams` to receive the wrong value.